+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - rename `McsArgs.repr` to `McsArgs.qualname`
 - add `McsArgs.is_abstract` property
 - make compatible with Python 3.5
+- ensure `abstract` meta option is always a boolean
 
 ## 0.3.0 (2018/09/30)